requirements.txt
robotframework==3.2.2
robotframework-requests==0.7.1
robotframework-seleniumlibrary==4.5.0
robotframework-sudslibrary-aljcalandra==1.1
requests==2.24.0
selenium==3.141.0
python-dateutil==2.8.1
robotframework-appiumlibrary==1.5.0.6
psutil==5.7.2
msedge-selenium-tools==3.141.2
